service_auth/requirements.txt
alembic==1.11.1
email-validator==2.0.0
# fastapi-jwt-auth==0.5.0
fastapi==0.95.2
fastapi-cache2[redis]==0.2.1
redis==4.5.5
SQLAlchemy==2.0.17
gunicorn==20.1.0
uvicorn==0.22.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
pydantic==1.10.7
psycopg[binary]==3.1.9
orjson==3.8.5
asyncpg==0.27.0
werkzeug==2.3.6
python-multipart==0.0.6
# email-validator==2.0.0.post2
bcrypt==4.0.1
passlib==1.7.4
# PyJWT==1.7.1
fastapi-jwt-auth[asymmetric]